+++ mayan/apps/authentication/forms.py
@@ -5,9 +5,8 @@ import warnings
 from django import forms
 from django.contrib.auth import authenticate
 from django.contrib.auth.forms import AuthenticationForm
+from django.forms.widgets import EmailInput
 from django.utils.translation import ugettext_lazy as _
-
-from mayan.apps.common.widgets import EmailInput
 
 
 class EmailAuthenticationForm(forms.Form):