+++ HISTORY.rst
@@ -162,6 +162,7 @@
 * Remove search elapsed time calculation.
 * Remove SEARCH_LIMIT setting.
 * Use the 'handler' prefix for all the signal handler functions.
+* Remove custom email widget and use Django's.
 
 3.1.11 (2019-04-XX)
 ===================